+++ src/ocean/util/serialize/model/VersionDecoratorMixins.d
@@ -18,6 +18,7 @@ module ocean.util.serialize.model.VersionDecoratorMixins;
 import ocean.transition;
 
 import ocean.core.Enforce,
+       ocean.core.Exception,
        ocean.util.container.ConcatBuffer,
        ocean.core.StructConverter;
 
@@ -424,16 +425,7 @@ unittest
 
 class VersionHandlingException : Exception
 {
-    /***************************************************************************
-
-        No-op constructor, all actual data gets set up by throwing methods
-
-    ***************************************************************************/
-
-    this()
-    {
-        super(null);
-    }
+    mixin ReusableExceptionImplementation;
 
     /***************************************************************************
 
@@ -455,13 +447,13 @@ class VersionHandlingException : Exception
     {
         if (input_length <= Version.Type.sizeof)
         {
-            this.msg = Format(
-                "Loading {} has failed, input buffer too short " ~
-                    "(length {}, need {})",
-                S.stringof,
-                input_length,
-                Version.Type.sizeof
-            );
+            this.set("Loading ")
+                .append(S.stringof)
+                .append(" has failed, input buffer too short (length ")
+                .append(input_length)
+                .append(", need ")
+                .append(Version.Type.sizeof)
+                .append(")");
             this.line = line;
             this.file = file;
 
@@ -487,12 +479,13 @@ class VersionHandlingException : Exception
     void throwCantConvert(S)(Version.Type input_version, istring file = __FILE__,
         int line = __LINE__)
     {
-        this.msg = Format(
-            "Got version {} for struct {}, expected {}. Can't convert between these",
-            input_version,
-            S.stringof,
-            Version.Info!(S).number
-        );
+        this.set("Got version ")
+            .append(input_version)
+            .append(" for struct ")
+            .append(S.stringof)
+            .append(", expected ")
+            .append(Version.Info!(S).number)
+            .append(". Can't convert between these");
         this.line = line;
         this.file = file;